Struct rusoto_health::AWSHealthClient

pub struct AWSHealthClient { /* fields omitted */ }

A client for the AWSHealth API.

Methods

impl AWSHealthClient

pub fn new(region: Region) -> AWSHealthClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> AWSHealthClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,