Struct rusoto_health::EventFilter

pub struct EventFilter {
    pub availability_zones: Option<Vec<String>>,
    pub end_times: Option<Vec<DateTimeRange>>,
    pub entity_arns: Option<Vec<String>>,
    pub entity_values: Option<Vec<String>>,
    pub event_arns: Option<Vec<String>>,
    pub event_status_codes: Option<Vec<String>>,
    pub event_type_categories: Option<Vec<String>>,
    pub event_type_codes: Option<Vec<String>>,
    pub last_updated_times: Option<Vec<DateTimeRange>>,
    pub regions: Option<Vec<String>>,
    pub services: Option<Vec<String>>,
    pub start_times: Option<Vec<DateTimeRange>>,
    pub tags: Option<Vec<HashMap<String, String>>>,
}

The values to use to filter results from the DescribeEvents and DescribeEventAggregates operations.

Fields
availability_zones: Option<Vec<String>>

A list of AWS availability zones.

end_times: Option<Vec<DateTimeRange>>

A list of dates and times that the event ended.

entity_arns: Option<Vec<String>>

A list of entity ARNs (unique identifiers).

entity_values: Option<Vec<String>>

A list of entity identifiers, such as EC2 instance IDs (i-34ab692e) or EBS volumes (vol-426ab23e).

event_arns: Option<Vec<String>>

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/AWS_EC2_MAINTENANCE_5331", "arn:aws:health:us-west-1::event/AWS_EBS_LOST_VOLUME_xyz"

event_status_codes: Option<Vec<String>>

A list of event status codes.

event_type_categories: Option<Vec<String>>

A list of event type category codes (issue, scheduledChange, or accountNotification).

event_type_codes: Option<Vec<String>>

A list of unique identifiers for event types. For example, "AWS_EC2_SYSTEM_MAINTENANCE_EVENT","AWS_RDS_MAINTENANCE_SCHEDULED"

last_updated_times: Option<Vec<DateTimeRange>>

A list of dates and times that the event was last updated.

regions: Option<Vec<String>>

A list of AWS regions.

services: Option<Vec<String>>

The AWS services associated with the event. For example, EC2, RDS.

start_times: Option<Vec<DateTimeRange>>

A list of dates and times that the event began.

tags: Option<Vec<HashMap<String, String>>>

A map of entity tags attached to the affected entity.