Trait rusoto_health::AWSHealth

pub trait AWSHealth {
    fn describe_affected_entities(
        &self, 
        input: DescribeAffectedEntitiesRequest
    ) -> RusotoFuture<DescribeAffectedEntitiesResponse, DescribeAffectedEntitiesError>;

    fn describe_entity_aggregates(
        &self, 
        input: DescribeEntityAggregatesRequest
    ) -> RusotoFuture<DescribeEntityAggregatesResponse, DescribeEntityAggregatesError>;

    fn describe_event_aggregates(
        &self, 
        input: DescribeEventAggregatesRequest
    ) -> RusotoFuture<DescribeEventAggregatesResponse, DescribeEventAggregatesError>;

    fn describe_event_details(
        &self, 
        input: DescribeEventDetailsRequest
    ) -> RusotoFuture<DescribeEventDetailsResponse, DescribeEventDetailsError>;

    fn describe_event_types(
        &self, 
        input: DescribeEventTypesRequest
    ) -> RusotoFuture<DescribeEventTypesResponse, DescribeEventTypesError>;

    fn describe_events(
        &self, 
        input: DescribeEventsRequest
    ) -> RusotoFuture<DescribeEventsResponse, DescribeEventsError>;
}

Trait representing the capabilities of the AWSHealth API. AWSHealth clients implement this trait.

Required methods

fn describe_affected_entities(
&self,
input: DescribeAffectedEntitiesRequest
) -> RusotoFuture<DescribeAffectedEntitiesResponse, DescribeAffectedEntitiesError>

Returns a list of entities that have been affected by the specified events, based on the specified filter criteria. Entities can refer to individual customer resources, groups of customer resources, or any other construct, depending on the AWS service. Events that have impact beyond that of the affected entities, or where the extent of impact is unknown, include at least one entity indicating this.

At least one event ARN is required. Results are sorted by the lastUpdatedTime of the entity, starting with the most recent.

fn describe_entity_aggregates(
&self,
input: DescribeEntityAggregatesRequest
) -> RusotoFuture<DescribeEntityAggregatesResponse, DescribeEntityAggregatesError>

Returns the number of entities that are affected by each of the specified events. If no events are specified, the counts of all affected entities are returned.

fn describe_event_aggregates(
&self,
input: DescribeEventAggregatesRequest
) -> RusotoFuture<DescribeEventAggregatesResponse, DescribeEventAggregatesError>

Returns the number of events of each event type (issue, scheduled change, and account notification). If no filter is specified, the counts of all events in each category are returned.

fn describe_event_details(
&self,
input: DescribeEventDetailsRequest
) -> RusotoFuture<DescribeEventDetailsResponse, DescribeEventDetailsError>

Returns detailed information about one or more specified events. Information includes standard event data (region, service, etc., as returned by DescribeEvents), a detailed event description, and possible additional metadata that depends upon the nature of the event. Affected entities are not included; to retrieve those, use the DescribeAffectedEntities operation.

If a specified event cannot be retrieved, an error message is returned for that event.

fn describe_event_types(
&self,
input: DescribeEventTypesRequest
) -> RusotoFuture<DescribeEventTypesResponse, DescribeEventTypesError>

Returns the event types that meet the specified filter criteria. If no filter criteria are specified, all event types are returned, in no particular order.

fn describe_events(
&self,
input: DescribeEventsRequest
) -> RusotoFuture<DescribeEventsResponse, DescribeEventsError>

Returns information about events that meet the specified filter criteria. Events are returned in a summary form and do not include the detailed description, any additional metadata that depends on the event type, or any affected resources. To retrieve that information, use the DescribeEventDetails and DescribeAffectedEntities operations.

If no filter criteria are specified, all events are returned. Results are sorted by lastModifiedTime, starting with the most recent.

Loading content...