Crate rusoto_health

AWS Health
The AWS Health API provides programmatic access to the AWS Health information that is presented in the AWS Personal Health Dashboard. You can get information about events that affect your AWS resources:

DescribeEvents: Summary information about events.

DescribeEventDetails: Detailed information about one or more events.

DescribeAffectedEntities: Information about AWS resources that are affected by one or more events.

In addition, these operations provide information about event types and summary counts of events or affected entities:

DescribeEventTypes: Information about the kinds of events that AWS Health tracks.

DescribeEventAggregates: A count of the number of events that meet specified criteria.

DescribeEntityAggregates: A count of the number of affected entities that meet specified criteria.

The Health API requires a Business or Enterprise support plan from AWS Support. Calling the Health API from an account that does not have a Business or Enterprise support plan causes a SubscriptionRequiredException.

For authentication of requests, AWS Health uses the Signature Version 4 Signing Process.

See the AWS Health User Guide for information about how to use the API.

Service Endpoint

The HTTP endpoint for the AWS Health API is:

https://health.us-east-1.amazonaws.com

If you're using the service, you're probably looking for AWSHealthClient and AWSHealth.

Structs

AWSHealthClient
A client for the AWSHealth API.

AffectedEntity
Information about an entity that is affected by a Health event.

DateTimeRange
A range of dates and times that is used by the EventFilter and EntityFilter objects. If from is set and to is set: match items where the timestamp (startTime, endTime, or lastUpdatedTime) is between from and to inclusive. If from is set and to is not set: match items where the timestamp value is equal to or after from. If from is not set and to is set: match items where the timestamp value is equal to or before to.

DescribeAffectedEntitiesRequest

DescribeAffectedEntitiesResponse

DescribeEntityAggregatesRequest

DescribeEntityAggregatesResponse

DescribeEventAggregatesRequest

DescribeEventAggregatesResponse

DescribeEventDetailsRequest

DescribeEventDetailsResponse

DescribeEventTypesRequest

DescribeEventTypesResponse

DescribeEventsRequest

DescribeEventsResponse

EntityAggregate
The number of entities that are affected by one or more events. Returned by the DescribeEntityAggregates operation.

EntityFilter
The values to use to filter results from the DescribeAffectedEntities operation.

Event
Summary information about an event, returned by the DescribeEvents operation. The DescribeEventDetails operation also returns this information, as well as the EventDescription and additional event metadata.

EventAggregate
The number of events of each issue type. Returned by the DescribeEventAggregates operation.

EventDetails
Detailed information about an event. A combination of an Event object, an EventDescription object, and additional metadata about the event. Returned by the DescribeEventDetails operation.

EventDetailsErrorItem
Error information returned when a DescribeEventDetails operation cannot find a specified event.

EventFilter
The values to use to filter results from the DescribeEvents and DescribeEventAggregates operations.

EventTypeFilter
The values to use to filter results from the DescribeEventTypes operation.

Enums

DescribeAffectedEntitiesError
Errors returned by DescribeAffectedEntities

DescribeEntityAggregatesError
Errors returned by DescribeEntityAggregates

DescribeEventAggregatesError
Errors returned by DescribeEventAggregates

DescribeEventDetailsError
Errors returned by DescribeEventDetails

DescribeEventTypesError
Errors returned by DescribeEventTypes

DescribeEventsError
Errors returned by DescribeEvents

Traits

AWSHealth
Trait representing the capabilities of the AWSHealth API. AWSHealth clients implement this trait.